refactor: rename typing/_typing.py to typing/numeric.py
The typing package's numeric-type and image-normalization module is
renamed to numeric.py to make room for an upcoming encoders.py protocol
module. The public API is unchanged: names are still re-exported from
pyvisim.typing.
