+++ pyvisim/typing/numeric.py
